test: split 12 more test files (admin/identity/economy/etc)
Continuation of file-splitting effort for flake reduction. Splits:
  test-helpers.test.js                   → -lifecycle, -misc
  admin-users-missing.test.js            → -core, -lifecycle, -moderation
  identity-graph.test.js                 → -read, -write
  admin-suggestions-extended.test.js     → -core, -merge
  admin-users.test.js                    → -read, -write
  admin-suggestions-new-routes.test.js   → -core, -dispute
  admin-economy.test.js                  → -read, -write
  suggestions-duplicates.test.js         → -core, -reject
  economy-coverage-gifts.test.js         → -core, -rewards, -send
  suggestions-contracts-a.test.js        → -edge, -misc
  suggestions-integration-a.test.js      → -submission, -flows
  reports-coverage-reports.test.js       → -core, -compat

Largest remaining file is now 1444 lines (was 3162). 10x flake check
maintains 60% clean rate; cluster failures further reduced (max 6
tests/run vs prior 50).
