+++ express-api/tests/routes/admin-suggestions-extended-core.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 /**
  * Extended tests for admin suggestion routes — gap coverage.
  *
@@ -640,196 +641,6 @@ describe('PUT /admin/suggestions/:id/link', () => {
 // ═══════════════════════════════════════════════════════════════
 // POST /admin/suggestions/:id/merge
 // ═══════════════════════════════════════════════════════════════
-
-describe('POST /admin/suggestions/:id/merge', () => {
-  test('already-merged suggestion returns 409', async () => {
-    setupDocMocks({
-      'suggestions/sug-dup-1': makeSuggestionSnap('sug-dup-1', {
-        status: 'merged',
-        mergedIntoSuggestionId: 'sug-orig-existing',
-      }),
-      'suggestions/sug-target': makeSuggestionSnap('sug-target', { status: 'accepted' }),
-    });
-
-    const app = createAdminApp();
-    // The merge endpoint uses mergeInto in the status route; the dedicated merge
-    // route marks as merged but does not check mergedIntoSuggestionId before.
-    // Test via the status endpoint with mergeInto parameter.
-    const res = await request(app)
-      .put('/api/admin/suggestions/sug-dup-1/status')
-      .send({ status: 'merged', mergeInto: 'sug-target' })
-      .expect(409);
-
-    expect(res.body.error).toMatch(/already merged/i);
-  });
-
-  test('self-merge (merging into itself) — merge endpoint does not prevent it explicitly', async () => {
-    // The dedicated merge endpoint (POST /merge) does not have a self-merge guard.
-    // This test documents the current behaviour: self-merge is NOT prevented.
-    // If the source and target are the same doc, both reads succeed.
-    setupDocMocks({
-      'suggestions/sug-self': makeSuggestionSnap('sug-self', {
-        status: 'accepted',
-        upvotes: 3,
-        submitterUid: 1001,
-      }),
-    });
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-self/merge')
-      .send({ originalSuggestionId: 'sug-self' });
-
-    // The endpoint processes it (no explicit self-merge guard exists).
-    // It will succeed with 200. This documents the gap.
-    expect([200, 400]).toContain(res.status);
-  });
-
-  test('non-admin returns 403', async () => {
-    const app = createApp({ uniqueId: 1001, isAdmin: false });
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-1/merge')
-      .send({ originalSuggestionId: 'sug-2' })
-      .expect(403);
-
-    expect(res.body.error).toMatch(/admin/i);
-  });
-
-  test('non-existent duplicate suggestion returns 404', async () => {
-    // mockDocGet defaults to { exists: false } for all paths
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/nonexistent/merge')
-      .send({ originalSuggestionId: 'sug-2' })
-      .expect(404);
-
-    expect(res.body.error).toMatch(/duplicate.*not found/i);
-  });
-
-  test('non-existent original/target suggestion returns 404', async () => {
-    setupDocMocks({
-      'suggestions/sug-dup-2': makeSuggestionSnap('sug-dup-2', { status: 'pending' }),
-      // Target does not exist — will fall through to { exists: false }
-    });
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-dup-2/merge')
-      .send({ originalSuggestionId: 'nonexistent-target' })
-      .expect(404);
-
-    expect(res.body.error).toMatch(/original.*not found/i);
-  });
-
-  test('missing target ID returns 400', async () => {
-    setupDocMocks({
-      'suggestions/sug-dup-3': makeSuggestionSnap('sug-dup-3', { status: 'pending' }),
-    });
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-dup-3/merge')
-      .send({})
-      .expect(400);
-
-    expect(res.body.error).toMatch(/original.*required/i);
-  });
-
-  test('successful merge sets status to merged and transfers votes', async () => {
-    setupDocMocks({
-      'suggestions/sug-dup-4': makeSuggestionSnap('sug-dup-4', {
-        status: 'pending',
-        upvotes: 7,
-        title: 'Duplicate feature',
-        submitterUid: 2002,
-      }),
-      'suggestions/sug-orig-4': makeSuggestionSnap('sug-orig-4', {
-        status: 'accepted',
-        upvotes: 10,
-      }),
-    });
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-dup-4/merge')
-      .send({ originalSuggestionId: 'sug-orig-4' })
-      .expect(200);
-
-    expect(res.body.success).toBe(true);
-
-    // Verify source updated to merged (db.doc(path).update(data))
-    expect(mockDocUpdate).toHaveBeenCalledWith(
-      expect.stringContaining('suggestions/sug-dup-4'),
-      expect.objectContaining({
-        status: 'merged',
-        mergedIntoSuggestionId: 'sug-orig-4',
-      }),
-    );
-
-    // Verify upvotes transferred to target
-    expect(mockDocUpdate).toHaveBeenCalledWith(
-      expect.stringContaining('suggestions/sug-orig-4'),
-      expect.objectContaining({
-        upvotes: expect.objectContaining({ _type: 'increment', value: 7 }),
-      }),
-    );
-
-    // Verify notification created for submitter
-    expect(mockCollectionAdd).toHaveBeenCalledWith(
-      'notifications',
-      expect.objectContaining({
-        uid: 2002,
-        type: 'suggestion_merged',
-      }),
-    );
-
-    // Verify audit log created
-    expect(mockCollectionAdd).toHaveBeenCalledWith(
-      'auditLog',
-      expect.objectContaining({
-        action: 'suggestion_merge',
-        details: expect.objectContaining({
-          duplicateId: 'sug-dup-4',
-          originalId: 'sug-orig-4',
-          transferredUpvotes: 7,
-        }),
-      }),
-    );
-  });
-
-  test('merge accepts targetId as alternative field name', async () => {
-    setupDocMocks({
-      'suggestions/sug-dup-5': makeSuggestionSnap('sug-dup-5', {
-        status: 'pending',
-        upvotes: 2,
-        submitterUid: 3003,
-      }),
-      'suggestions/sug-orig-5': makeSuggestionSnap('sug-orig-5', { status: 'accepted' }),
-    });
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-dup-5/merge')
-      .send({ targetId: 'sug-orig-5' })
-      .expect(200);
-
-    expect(res.body.success).toBe(true);
-  });
-
-  test('merge accepts targetSuggestionId as alternative field name', async () => {
-    setupDocMocks({
-      'suggestions/sug-dup-6': makeSuggestionSnap('sug-dup-6', {
-        status: 'pending',
-        upvotes: 1,
-        submitterUid: 4004,
-      }),
-      'suggestions/sug-orig-6': makeSuggestionSnap('sug-orig-6', { status: 'accepted' }),
-    });
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/admin/suggestions/sug-dup-6/merge')
-      .send({ targetSuggestionId: 'sug-orig-6' })
-      .expect(200);
-
-    expect(res.body.success).toBe(true);
-  });
-});
 
 // ═══════════════════════════════════════════════════════════════
 // DELETE /admin/suggestions/blocked/:id
