+++ express-api/tests/routes/admin-suggestions-extended-merge.test.js
@@ -1,0 +1,496 @@
+/* eslint-disable no-unused-vars */
+/**
+ * Extended tests for admin suggestion routes — gap coverage.
+ *
+ * Covers:
+ *   - Invalid status transitions (all disallowed pairs)
+ *   - Blocked topic creation on rejection
+ *   - Blocked topic cleanup when overturning rejection (rejected → accepted)
+ *   - PUT /admin/suggestions/:id/link edge cases (non-admin, non-existent)
+ *   - POST /admin/suggestions/:id/merge edge cases (already merged, self-merge, non-admin)
+ *   - DELETE /admin/suggestions/blocked/:id edge cases (non-existent, non-admin)
+ *   - GET /admin/suggestions/disputes edge cases (non-admin, empty)
+ *   - PUT /admin/suggestions/disputes/:id edge cases (invalid resolution)
+ *   - POST /suggestions/:id/dispute edge cases (unauthenticated, non-submitter, already disputed)
+ *
+ * Routes under test:
+ *   PUT    /api/admin/suggestions/:id/status
+ *   PUT    /api/admin/suggestions/:id/link
+ *   POST   /api/admin/suggestions/:id/merge
+ *   DELETE /api/admin/suggestions/blocked/:id
+ *   GET    /api/admin/suggestions/disputes
+ *   PUT    /api/admin/suggestions/disputes/:id
+ *   POST   /api/suggestions/:id/dispute
+ */
+
+const express = require('express');
+const request = require('supertest');
+
+// ─── Firebase mock ──────────────────────────────────────────────
+
+const mockDocGet = jest.fn();
+const mockDocSet = jest.fn().mockResolvedValue();
+const mockDocUpdate = jest.fn().mockResolvedValue();
+const mockDocDelete = jest.fn().mockResolvedValue();
+
+const mockCollectionAdd = jest.fn().mockResolvedValue({ id: 'new-id' });
+const mockCollectionGet = jest.fn().mockResolvedValue({ empty: true, docs: [], size: 0 });
+
+const mockWhere = jest.fn();
+const mockOrderBy = jest.fn();
+const mockLimit = jest.fn();
+
+const mockQueryChain = {
+  where: (...args) => {
+    mockWhere(...args);
+    return mockQueryChain;
+  },
+  orderBy: (...args) => {
+    mockOrderBy(...args);
+    return mockQueryChain;
+  },
+  limit: (...args) => {
+    mockLimit(...args);
+    return mockQueryChain;
+  },
+  get: () => mockCollectionGet(),
+};
+
+const mockRunTransaction = jest.fn(async (fn) => {
+  const t = {
+    get: mockDocGet,
+    set: mockDocSet,
+    update: mockDocUpdate,
+    delete: mockDocDelete,
+  };
+  return fn(t);
+});
+
+jest.mock('../../src/utils/firebase', () => ({
+  db: {
+    doc: jest.fn((path) => ({
+      _path: path,
+      get: () => mockDocGet(path),
+      set: (...args) => mockDocSet(path, ...args),
+      update: (...args) => mockDocUpdate(path, ...args),
+      delete: () => mockDocDelete(path),
+    })),
+    collection: jest.fn((name) => ({
+      _name: name,
+      add: (...args) => mockCollectionAdd(name, ...args),
+      doc: jest.fn((id) => ({
+        get: () => mockDocGet(`${name}/${id}`),
+        set: (...args) => mockDocSet(`${name}/${id}`, ...args),
+        update: (...args) => mockDocUpdate(`${name}/${id}`, ...args),
+        delete: () => mockDocDelete(`${name}/${id}`),
+      })),
+      where: (...args) => {
+        mockWhere(...args);
+        return mockQueryChain;
+      },
+      orderBy: (...args) => {
+        mockOrderBy(...args);
+        return mockQueryChain;
+      },
+      get: () => mockCollectionGet(),
+    })),
+    runTransaction: mockRunTransaction,
+  },
+  FieldValue: {
+    serverTimestamp: jest.fn(() => 'SERVER_TIMESTAMP'),
+    arrayUnion: jest.fn((...args) => ({ _type: 'arrayUnion', values: args })),
+    arrayRemove: jest.fn((...args) => ({ _type: 'arrayRemove', values: args })),
+    increment: jest.fn((n) => ({ _type: 'increment', value: n })),
+    delete: jest.fn(() => ({ _type: 'delete' })),
+  },
+}));
+
+jest.mock('../../src/utils/helpers', () => ({
+  generateId: jest.fn(() => 'mock-id-123'),
+  now: jest.fn(() => 1709913600000),
+}));
+
+jest.mock('../../src/utils/log', () => ({
+  debug: jest.fn(),
+  info: jest.fn(),
+  warn: jest.fn(),
+  error: jest.fn(),
+}));
+
+jest.mock('../../src/utils/system-pm', () => ({
+  sendSystemPm: jest.fn().mockResolvedValue(),
+}));
+
+jest.mock('../../src/utils/fcm', () => ({
+  sendFcmToTokens: jest.fn().mockResolvedValue([]),
+}));
+
+jest.mock('../../src/utils/roadmap-notify', () => ({
+  notifyRoadmapSubscribers: jest.fn().mockResolvedValue(),
+}));
+
+// ─── App setup ──────────────────────────────────────────────────
+
+const suggestionsRouter = require('../../src/routes/suggestions');
+
+function createApp({ uniqueId = 1001, isAdmin = false, isSuspended = false } = {}) {
+  const app = express();
+  app.use(express.json());
+  app.use((req, _res, next) => {
+    req.auth = {
+      uid: `firebase-uid-${uniqueId}`,
+      uniqueId,
+      token: { admin: isAdmin },
+    };
+    if (isSuspended) {
+      req.auth.suspended = true;
+    }
+    next();
+  });
+  app.use('/api', suggestionsRouter);
+  return app;
+}
+
+function createAdminApp({ uniqueId = 'admin-1' } = {}) {
+  return createApp({ uniqueId, isAdmin: true });
+}
+
+function createUnauthenticatedApp() {
+  const app = express();
+  app.use(express.json());
+  app.use((req, _res, next) => {
+    // No auth at all
+    req.auth = null;
+    next();
+  });
+  app.use('/api', suggestionsRouter);
+  return app;
+}
+
+beforeEach(() => {
+  jest.clearAllMocks();
+  mockDocGet.mockReset();
+  mockCollectionGet.mockReset();
+  mockRunTransaction.mockReset();
+  mockRunTransaction.mockImplementation(async (fn) => {
+    const t = { get: mockDocGet, set: mockDocSet, update: mockDocUpdate, delete: mockDocDelete };
+    return fn(t);
+  });
+  mockDocGet.mockResolvedValue({ exists: false });
+  mockCollectionGet.mockResolvedValue({ empty: true, docs: [], size: 0 });
+});
+
+// ─── Helpers ────────────────────────────────────────────────────
+
+function makeSuggestionSnap(id, overrides = {}) {
+  return {
+    id,
+    exists: true,
+    data: () => ({
+      title: 'Test suggestion',
+      description: 'A test description that is valid',
+      tags: ['quality-of-life'],
+      language: 'en',
+      status: 'pending',
+      rejectReason: null,
+      linkedRoadmapFeature: null,
+      mergedIntoSuggestionId: null,
+      disputePending: false,
+      submitterUid: 1001,
+      submitterContactOptIn: false,
+      upvotes: 5,
+      downvotes: 1,
+      createdAt: 1709913600000,
+      updatedAt: 1709913600000,
+      reviewedAt: null,
+      reviewedBy: null,
+      completedAt: null,
+      editHistory: [],
+      subscribers: [1001],
+      votingLocked: false,
+      commentsLocked: false,
+      ...overrides,
+    }),
+  };
+}
+
+function makeDisputeSnap(id, overrides = {}) {
+  return {
+    id,
+    exists: true,
+    data: () => ({
+      suggestionId: 'sug-merged-1',
+      originalSuggestionId: 'sug-original-1',
+      submitterUid: 1001,
+      reason: 'These are different features',
+      status: 'pending',
+      createdAt: 1709913600000,
+      resolvedAt: null,
+      resolvedBy: null,
+      resolution: null,
+      ...overrides,
+    }),
+  };
+}
+
+function makeRoadmapFeatureSnap(id, overrides = {}) {
+  return {
+    id,
+    exists: true,
+    data: () => ({
+      title: 'Roadmap feature',
+      phase: 'phase-2',
+      status: 'planned',
+      ...overrides,
+    }),
+  };
+}
+
+function makeBlockedTopicSnap(id, overrides = {}) {
+  return {
+    id,
+    exists: true,
+    data: () => ({
+      title: 'Blocked topic title',
+      reason: 'Already planned internally',
+      rejectReason: 'Already planned internally',
+      originalSuggestionId: 'sug-orig-1',
+      createdAt: 1709913600000,
+      ...overrides,
+    }),
+  };
+}
+
+/** Set up mockDocGet to resolve different docs based on path. */
+function setupDocMocks(pathMap) {
+  mockDocGet.mockImplementation((path) => {
+    for (const [pattern, snap] of Object.entries(pathMap)) {
+      if (typeof path === 'string' && path.includes(pattern)) {
+        return Promise.resolve(snap);
+      }
+    }
+    return Promise.resolve({ exists: false });
+  });
+}
+
+// ═══════════════════════════════════════════════════════════════
+// PUT /admin/suggestions/:id/status — Invalid transitions
+// ═══════════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════════
+// PUT /admin/suggestions/:id/status — Blocked topic side effects
+// ═══════════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════════
+// PUT /admin/suggestions/:id/link
+// ═══════════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════════
+// POST /admin/suggestions/:id/merge
+// ═══════════════════════════════════════════════════════════════
+
+describe('POST /admin/suggestions/:id/merge', () => {
+  test('already-merged suggestion returns 409', async () => {
+    setupDocMocks({
+      'suggestions/sug-dup-1': makeSuggestionSnap('sug-dup-1', {
+        status: 'merged',
+        mergedIntoSuggestionId: 'sug-orig-existing',
+      }),
+      'suggestions/sug-target': makeSuggestionSnap('sug-target', { status: 'accepted' }),
+    });
+
+    const app = createAdminApp();
+    // The merge endpoint uses mergeInto in the status route; the dedicated merge
+    // route marks as merged but does not check mergedIntoSuggestionId before.
+    // Test via the status endpoint with mergeInto parameter.
+    const res = await request(app)
+      .put('/api/admin/suggestions/sug-dup-1/status')
+      .send({ status: 'merged', mergeInto: 'sug-target' })
+      .expect(409);
+
+    expect(res.body.error).toMatch(/already merged/i);
+  });
+
+  test('self-merge (merging into itself) — merge endpoint does not prevent it explicitly', async () => {
+    // The dedicated merge endpoint (POST /merge) does not have a self-merge guard.
+    // This test documents the current behaviour: self-merge is NOT prevented.
+    // If the source and target are the same doc, both reads succeed.
+    setupDocMocks({
+      'suggestions/sug-self': makeSuggestionSnap('sug-self', {
+        status: 'accepted',
+        upvotes: 3,
+        submitterUid: 1001,
+      }),
+    });
+
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-self/merge')
+      .send({ originalSuggestionId: 'sug-self' });
+
+    // The endpoint processes it (no explicit self-merge guard exists).
+    // It will succeed with 200. This documents the gap.
+    expect([200, 400]).toContain(res.status);
+  });
+
+  test('non-admin returns 403', async () => {
+    const app = createApp({ uniqueId: 1001, isAdmin: false });
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-1/merge')
+      .send({ originalSuggestionId: 'sug-2' })
+      .expect(403);
+
+    expect(res.body.error).toMatch(/admin/i);
+  });
+
+  test('non-existent duplicate suggestion returns 404', async () => {
+    // mockDocGet defaults to { exists: false } for all paths
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/nonexistent/merge')
+      .send({ originalSuggestionId: 'sug-2' })
+      .expect(404);
+
+    expect(res.body.error).toMatch(/duplicate.*not found/i);
+  });
+
+  test('non-existent original/target suggestion returns 404', async () => {
+    setupDocMocks({
+      'suggestions/sug-dup-2': makeSuggestionSnap('sug-dup-2', { status: 'pending' }),
+      // Target does not exist — will fall through to { exists: false }
+    });
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-dup-2/merge')
+      .send({ originalSuggestionId: 'nonexistent-target' })
+      .expect(404);
+
+    expect(res.body.error).toMatch(/original.*not found/i);
+  });
+
+  test('missing target ID returns 400', async () => {
+    setupDocMocks({
+      'suggestions/sug-dup-3': makeSuggestionSnap('sug-dup-3', { status: 'pending' }),
+    });
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-dup-3/merge')
+      .send({})
+      .expect(400);
+
+    expect(res.body.error).toMatch(/original.*required/i);
+  });
+
+  test('successful merge sets status to merged and transfers votes', async () => {
+    setupDocMocks({
+      'suggestions/sug-dup-4': makeSuggestionSnap('sug-dup-4', {
+        status: 'pending',
+        upvotes: 7,
+        title: 'Duplicate feature',
+        submitterUid: 2002,
+      }),
+      'suggestions/sug-orig-4': makeSuggestionSnap('sug-orig-4', {
+        status: 'accepted',
+        upvotes: 10,
+      }),
+    });
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-dup-4/merge')
+      .send({ originalSuggestionId: 'sug-orig-4' })
+      .expect(200);
+
+    expect(res.body.success).toBe(true);
+
+    // Verify source updated to merged (db.doc(path).update(data))
+    expect(mockDocUpdate).toHaveBeenCalledWith(
+      expect.stringContaining('suggestions/sug-dup-4'),
+      expect.objectContaining({
+        status: 'merged',
+        mergedIntoSuggestionId: 'sug-orig-4',
+      }),
+    );
+
+    // Verify upvotes transferred to target
+    expect(mockDocUpdate).toHaveBeenCalledWith(
+      expect.stringContaining('suggestions/sug-orig-4'),
+      expect.objectContaining({
+        upvotes: expect.objectContaining({ _type: 'increment', value: 7 }),
+      }),
+    );
+
+    // Verify notification created for submitter
+    expect(mockCollectionAdd).toHaveBeenCalledWith(
+      'notifications',
+      expect.objectContaining({
+        uid: 2002,
+        type: 'suggestion_merged',
+      }),
+    );
+
+    // Verify audit log created
+    expect(mockCollectionAdd).toHaveBeenCalledWith(
+      'auditLog',
+      expect.objectContaining({
+        action: 'suggestion_merge',
+        details: expect.objectContaining({
+          duplicateId: 'sug-dup-4',
+          originalId: 'sug-orig-4',
+          transferredUpvotes: 7,
+        }),
+      }),
+    );
+  });
+
+  test('merge accepts targetId as alternative field name', async () => {
+    setupDocMocks({
+      'suggestions/sug-dup-5': makeSuggestionSnap('sug-dup-5', {
+        status: 'pending',
+        upvotes: 2,
+        submitterUid: 3003,
+      }),
+      'suggestions/sug-orig-5': makeSuggestionSnap('sug-orig-5', { status: 'accepted' }),
+    });
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-dup-5/merge')
+      .send({ targetId: 'sug-orig-5' })
+      .expect(200);
+
+    expect(res.body.success).toBe(true);
+  });
+
+  test('merge accepts targetSuggestionId as alternative field name', async () => {
+    setupDocMocks({
+      'suggestions/sug-dup-6': makeSuggestionSnap('sug-dup-6', {
+        status: 'pending',
+        upvotes: 1,
+        submitterUid: 4004,
+      }),
+      'suggestions/sug-orig-6': makeSuggestionSnap('sug-orig-6', { status: 'accepted' }),
+    });
+    const app = createAdminApp();
+    const res = await request(app)
+      .post('/api/admin/suggestions/sug-dup-6/merge')
+      .send({ targetSuggestionId: 'sug-orig-6' })
+      .expect(200);
+
+    expect(res.body.success).toBe(true);
+  });
+});
+
+// ═══════════════════════════════════════════════════════════════
+// DELETE /admin/suggestions/blocked/:id
+// ═══════════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════════
+// GET /admin/suggestions/disputes
+// ═══════════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════════
+// PUT /admin/suggestions/disputes/:id
+// ═══════════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════════
+// POST /suggestions/:id/dispute
+// ═══════════════════════════════════════════════════════════════
