+++ express-api/tests/routes/economy-coverage-gifts-rewards.test.js
@@ -219,93 +219,9 @@ function makeBackpackDoc(quantity = 5) {
 // updateGiftWall — lines 134-135 (existing sender increment)
 // ═══════════════════════════════════════════════════════════════════
 
-describe('updateGiftWall — existing sender (lines 134-135)', () => {
-  test('increments existing sender sendCount in gift wall', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 }))
-      .mockResolvedValueOnce(makeBackpackDoc(5))
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500, displayName: 'Alice' }))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({
-        exists: true,
-        data: () => ({
-          giftId: 'gift-rose',
-          receivedCount: 3,
-          senders: [{ senderId: 'user-A', sendCount: 3, lastSentAt: 1700000000000 }],
-        }),
-      })
-      .mockResolvedValueOnce({
-        exists: true,
-        data: () => ({
-          rankings: [{ userId: 'user-B', count: 5, rank: 1 }],
-          totalSent: 5,
-        }),
-      })
-      .mockResolvedValue({ exists: false });
-
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 5 }) }),
-        update: jest.fn(),
-        delete: jest.fn(),
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(200);
-    expect(res.body.success).toBe(true);
-  });
-});
-
 // ═══════════════════════════════════════════════════════════════════
 // updateGiftRankings — line 194 (existing ranking entry increment)
 // ═══════════════════════════════════════════════════════════════════
-
-describe('updateGiftRankings — existing user (line 194)', () => {
-  test('increments existing ranking count for recipient', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 }))
-      .mockResolvedValueOnce(makeBackpackDoc(5))
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({
-        exists: true,
-        data: () => ({
-          rankings: [
-            { userId: 'user-B', count: 10, rank: 1, displayName: 'Bob' },
-            { userId: 'user-C', count: 5, rank: 2, displayName: 'Carol' },
-          ],
-          totalSent: 15,
-        }),
-      })
-      .mockResolvedValue({ exists: false });
-
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 5 }) }),
-        update: jest.fn(),
-        delete: jest.fn(),
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(200);
-    expect(mockDocSet).toHaveBeenCalled();
-  });
-});
 
 // ═══════════════════════════════════════════════════════════════════
 // Daily reward — lines 259 (gift reward milestone), 283-285, 308-309
@@ -762,255 +678,10 @@ describe('POST /api/economy/gacha — advanced mechanics', () => {
 // 740 (broadcast), 757-758 (500 error)
 // ═══════════════════════════════════════════════════════════════════
 
-describe('POST /api/economy/gift — additional coverage', () => {
-  test('returns 400 for super_shy_trial gift (line 616)', async () => {
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'super_shy_trial' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/trial items cannot be transferred/i);
-  });
-
-  test('returns 400 when sending gift to yourself (line 618)', async () => {
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-A', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/cannot send gift to yourself/i);
-  });
-
-  test('deletes backpack item when quantity reaches 0 (line 665-666)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc())
-      .mockResolvedValueOnce(makeBackpackDoc(1))
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    const deleteMock = jest.fn();
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 1 }) }),
-        update: jest.fn(),
-        delete: deleteMock,
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(200);
-    expect(deleteMock).toHaveBeenCalled();
-  });
-
-  test('writes room gift message when sender is in a room (lines 681-694)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc())
-      .mockResolvedValueOnce(makeBackpackDoc(5))
-      .mockResolvedValueOnce(makeUserDoc({ currentRoomId: 'room-1', displayName: 'Alice' }))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 5 }) }),
-        update: jest.fn(),
-        delete: jest.fn(),
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(200);
-    expect(mockDocSet).toHaveBeenCalled();
-    expect(mockDocUpdate).toHaveBeenCalled();
-  });
-
-  test('writes room gift message with quantity label when qty > 1 (line 683)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc())
-      .mockResolvedValueOnce(makeBackpackDoc(5))
-      .mockResolvedValueOnce(makeUserDoc({ currentRoomId: 'room-1', displayName: 'Alice' }))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 5 }) }),
-        update: jest.fn(),
-        delete: jest.fn(),
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 3 });
-
-    expect(res.status).toBe(200);
-    expect(res.body.quantity).toBe(3);
-  });
-
-  test('returns 500 on internal error (lines 757-758)', async () => {
-    mockDocGet.mockRejectedValue(new Error('Firestore down'));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(500);
-    expect(log.error).toHaveBeenCalledWith(
-      'economy',
-      expect.stringContaining('/economy/gift'),
-      expect.any(Object),
-    );
-  });
-});
-
 // ═══════════════════════════════════════════════════════════════════
 // Gift-direct — lines 771,773 (validation), 795-799 (block),
 // 825-828 (room message), 869 (broadcast), 885-886 (500 error)
 // ═══════════════════════════════════════════════════════════════════
-
-describe('POST /api/economy/gift-direct — additional coverage', () => {
-  test('returns 400 when recipientId is missing (line 771)', async () => {
-    const app = createApp('user-A');
-    const res = await request(app).post('/api/economy/gift-direct').send({ giftId: 'gift-rose' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/recipientId/i);
-  });
-
-  test('returns 400 when sending to yourself (line 773)', async () => {
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-A', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/cannot send gift to yourself/i);
-  });
-
-  test('returns 404 when gift not found', async () => {
-    mockDocGet
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce(makeUserDoc());
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/gift not found/i);
-  });
-
-  test('returns 404 when recipient not found', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc())
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce({ exists: false });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/recipient not found/i);
-  });
-
-  test('returns 403 when sender blocked recipient (lines 795-799)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc())
-      .mockResolvedValueOnce(makeUserDoc({ blockedUserIds: ['user-B'] }))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(403);
-    expect(res.body.error).toMatch(/blocked users/i);
-    expect(log.warn).toHaveBeenCalled();
-  });
-
-  test('returns 403 when recipient blocked sender', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc())
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob', blockedUserIds: ['user-A'] }));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(403);
-    expect(res.body.error).toMatch(/blocked users/i);
-  });
-
-  test('writes room message when sender is in a room (lines 825-828)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 }))
-      .mockResolvedValueOnce(
-        makeUserDoc({ shyCoins: 500, currentRoomId: 'room-1', displayName: 'Alice' }),
-      )
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(200);
-    expect(mockDocSet).toHaveBeenCalled();
-  });
-
-  test('returns 500 on internal error (lines 885-886)', async () => {
-    mockDocGet.mockRejectedValue(new Error('Firestore down'));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose' });
-
-    expect(res.status).toBe(500);
-    expect(log.error).toHaveBeenCalledWith(
-      'economy',
-      expect.stringContaining('gift-direct'),
-      expect.any(Object),
-    );
-  });
-});
 
 // ═══════════════════════════════════════════════════════════════════
 // Gift-batch — lines 902,904 (trial/self), 921-924 (insufficient backpack),
@@ -1018,388 +689,10 @@ describe('POST /api/economy/gift-direct — additional coverage', () => {
 // 1025-1026 (room message), 1056-1057 (500 error)
 // ═══════════════════════════════════════════════════════════════════
 
-describe('POST /api/economy/gift-batch — additional coverage', () => {
-  test('returns 400 for super_shy_trial gift (line 902)', async () => {
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'super_shy_trial' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/trial items cannot be transferred/i);
-  });
-
-  test('returns 400 when sender is in recipientIds (line 904)', async () => {
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-A', 'user-B'], giftId: 'gift-rose' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/cannot send gift to yourself/i);
-  });
-
-  test('returns 404 when gift not found', async () => {
-    mockDocGet.mockResolvedValueOnce({ exists: false });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose' });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/gift not found/i);
-  });
-
-  test('returns 404 when sender not found', async () => {
-    mockDocGet.mockResolvedValueOnce(makeGiftDoc()).mockResolvedValueOnce({ exists: false });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose' });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/sender not found/i);
-  });
-
-  test('returns 402 when backpack has insufficient items (lines 921-924)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 }))
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500 }))
-      .mockResolvedValueOnce(makeBackpackDoc(1));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({
-        recipientIds: ['user-B', 'user-C'],
-        giftId: 'gift-rose',
-        quantity: 1,
-        fromBackpack: true,
-      });
-
-    expect(res.status).toBe(402);
-    expect(res.body.error).toMatch(/insufficient items in backpack/i);
-  });
-
-  test('returns 404 when no valid recipients exist', async () => {
-    // gift-batch flow: gift -> sender -> (coins check) -> loadEconomyConfig -> recipients
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 })) // gift
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500 })) // sender
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC) // loadEconomyConfig
-      // All recipients don't exist
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B', 'user-C'], giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/no valid recipients/i);
-  });
-
-  test('returns 403 when one recipient has blocked sender (lines 951-955)', async () => {
-    // gift-batch flow: gift -> sender -> (coins check) -> loadEconomyConfig -> recipients -> block check
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 })) // gift
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500 })) // sender
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC) // loadEconomyConfig
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob', blockedUserIds: ['user-A'] })); // recipient blocks sender
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(403);
-    expect(res.body.error).toMatch(/blocked users/i);
-  });
-
-  test('returns 403 when sender has blocked a recipient', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 })) // gift
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500, blockedUserIds: ['user-B'] })) // sender blocks user-B
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC) // loadEconomyConfig
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' })); // recipient
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(403);
-    expect(res.body.error).toMatch(/blocked users/i);
-  });
-
-  test('fromBackpack mode deducts from backpack via transaction (lines 960-977)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 }))
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500 }))
-      .mockResolvedValueOnce(makeBackpackDoc(5))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    const txDeleteMock = jest.fn();
-    const txUpdateMock = jest.fn();
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 5 }) }),
-        update: txUpdateMock,
-        delete: txDeleteMock,
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose', quantity: 1, fromBackpack: true });
-
-    expect(res.status).toBe(200);
-    expect(txUpdateMock).toHaveBeenCalled();
-  });
-
-  test('batch with room message when sender is in room (lines 1025-1026)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 }))
-      .mockResolvedValueOnce(
-        makeUserDoc({ shyCoins: 500, currentRoomId: 'room-1', displayName: 'Alice' }),
-      )
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ shyCoins: 500 }) }),
-        update: jest.fn(),
-        delete: jest.fn(),
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose', quantity: 1, fromBackpack: false });
-
-    expect(res.status).toBe(200);
-    expect(mockDocSet).toHaveBeenCalled();
-  });
-
-  test('returns 500 on internal error (lines 1056-1057)', async () => {
-    mockDocGet.mockRejectedValue(new Error('Firestore down'));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose' });
-
-    expect(res.status).toBe(500);
-    expect(log.error).toHaveBeenCalledWith(
-      'economy',
-      expect.stringContaining('gift-batch'),
-      expect.any(Object),
-    );
-  });
-});
-
 // ═══════════════════════════════════════════════════════════════════
 // Backpack-send — lines 1087-1091 (block), 1117-1118 (coinValue lookup),
 // 1176 (room message), 1188-1189 (500 error)
 // ═══════════════════════════════════════════════════════════════════
-
-describe('POST /api/economy/backpack-send — additional coverage', () => {
-  test('returns 400 when recipientId is missing', async () => {
-    const app = createApp('user-A');
-    const res = await request(app).post('/api/economy/backpack-send').send({});
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/recipientId/i);
-  });
-
-  test('returns 400 when sending to yourself', async () => {
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-A' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/cannot send to yourself/i);
-  });
-
-  test('returns 404 when sender not found', async () => {
-    mockDocGet.mockResolvedValueOnce({ exists: false }).mockResolvedValueOnce(makeUserDoc());
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/sender not found/i);
-  });
-
-  test('returns 404 when recipient not found', async () => {
-    mockDocGet.mockResolvedValueOnce(makeUserDoc()).mockResolvedValueOnce({ exists: false });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/recipient not found/i);
-  });
-
-  test('returns 403 when sender blocked recipient (lines 1087-1091)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeUserDoc({ blockedUserIds: ['user-B'] }))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(403);
-    expect(res.body.error).toMatch(/blocked users/i);
-    expect(log.warn).toHaveBeenCalled();
-  });
-
-  test('returns 403 when recipient blocked sender', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob', blockedUserIds: ['user-A'] }));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(403);
-    expect(res.body.error).toMatch(/blocked users/i);
-  });
-
-  test('looks up gift coinValue when not denormalized on backpack doc (lines 1117-1118)', async () => {
-    // backpack-send flow: sender -> recipient -> backpack(collection) -> loadEconomyConfig
-    //   -> for each item: gift lookup -> updateGiftWall -> updateGiftRankings
-    mockDocGet
-      .mockResolvedValueOnce(makeUserDoc()) // 1. sender
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' })) // 2. recipient
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC) // 3. loadEconomyConfig
-      // 4. Gift catalog lookup (coinValue missing from backpack doc)
-      .mockResolvedValueOnce({
-        exists: true,
-        data: () => ({ coinValue: 50 }),
-      })
-      // 5. updateGiftWall - giftWall doc
-      .mockResolvedValueOnce({ exists: false })
-      // 6. updateGiftRankings - giftRankings doc
-      .mockResolvedValueOnce({ exists: false })
-      // 7. updateGiftRankings - user doc (for new ranking entry)
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      // remaining calls
-      .mockResolvedValue({ exists: false });
-
-    mockCollectionGet = jest.fn().mockResolvedValue({
-      empty: false,
-      docs: [
-        {
-          id: 'gift-rare',
-          data: () => ({ giftId: 'gift-rare', quantity: 2 }),
-        },
-      ],
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(200);
-    expect(res.body.success).toBe(true);
-    // Beans = floor(50 * 0.6 * 2) = 60
-    expect(res.body.totalBeanReward).toBe(60);
-  });
-
-  test('writes room message when sender is in a room (line 1176)', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeUserDoc({ currentRoomId: 'room-1', displayName: 'Alice' }))
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }))
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC)
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValueOnce({ exists: false })
-      .mockResolvedValue({ exists: false });
-
-    mockCollectionGet = jest.fn().mockResolvedValue({
-      empty: false,
-      docs: [
-        {
-          id: 'gift-rose',
-          data: () => ({ giftId: 'gift-rose', quantity: 1, coinValue: 10 }),
-        },
-      ],
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(200);
-    expect(mockDocSet).toHaveBeenCalled();
-  });
-
-  test('filters out super_shy_trial items from backpack send', async () => {
-    mockDocGet
-      .mockResolvedValueOnce(makeUserDoc())
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' }));
-
-    mockCollectionGet = jest.fn().mockResolvedValue({
-      empty: false,
-      docs: [
-        {
-          id: 'super_shy_trial',
-          data: () => ({ giftId: 'super_shy_trial', quantity: 1, coinValue: 0 }),
-        },
-      ],
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/backpack is empty/i);
-  });
-
-  test('returns 500 on internal error (lines 1188-1189)', async () => {
-    mockDocGet.mockRejectedValue(new Error('Firestore down'));
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/backpack-send')
-      .send({ recipientId: 'user-B' });
-
-    expect(res.status).toBe(500);
-    expect(log.error).toHaveBeenCalledWith(
-      'economy',
-      expect.stringContaining('backpack-send'),
-      expect.any(Object),
-    );
-  });
-});
 
 // ═══════════════════════════════════════════════════════════════════
 // Redeem beans — lines 1235-1236 (500 error)
@@ -1449,38 +742,6 @@ describe('POST /api/economy/backpack-send — additional coverage', () => {
 // ═══════════════════════════════════════════════════════════════════
 // updateGiftRankings error path — line 222
 // ═══════════════════════════════════════════════════════════════════
-
-describe('updateGiftRankings error handling (line 222)', () => {
-  test('catches and logs error when ranking update fails', async () => {
-    // Make the giftRankings doc.get() throw during updateGiftRankings
-    // The gift-direct route calls updateGiftRankings near the end.
-    // We need all mocks up to that point to succeed, then fail on the rankings doc.
-    let callCount = 0;
-    mockDocGet.mockImplementation(() => {
-      callCount++;
-      if (callCount === 1) return Promise.resolve(makeGiftDoc({ coinValue: 10 })); // gift
-      if (callCount === 2) return Promise.resolve(makeUserDoc({ shyCoins: 500 })); // sender
-      if (callCount === 3) return Promise.resolve(makeUserDoc({ displayName: 'Bob' })); // recipient
-      if (callCount === 4) return Promise.resolve(ECONOMY_CONFIG_DOC); // config
-      if (callCount === 5) return Promise.resolve({ exists: false }); // giftWall doc (new)
-      if (callCount === 6) return Promise.reject(new Error('Rankings fetch failed')); // giftRankings
-      return Promise.resolve({ exists: false });
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-direct')
-      .send({ recipientId: 'user-B', giftId: 'gift-rose', quantity: 1 });
-
-    expect(res.status).toBe(200);
-    // The error should have been caught internally and logged
-    expect(log.error).toHaveBeenCalledWith(
-      'economy',
-      expect.stringContaining('gift rankings'),
-      expect.any(Object),
-    );
-  });
-});
 
 // ═══════════════════════════════════════════════════════════════════
 // Gacha broadcast error — line 582
@@ -1563,40 +824,6 @@ describe('POST /api/economy/gacha — broadcast error (line 582)', () => {
 // ═══════════════════════════════════════════════════════════════════
 // Gift-batch fromBackpack delete path — line 967
 // ═══════════════════════════════════════════════════════════════════
-
-describe('POST /api/economy/gift-batch — backpack delete when qty reaches 0 (line 967)', () => {
-  test('deletes backpack item when remaining quantity is 0', async () => {
-    // Send 1 gift to 1 recipient from backpack, backpack has exactly 1
-    mockDocGet
-      .mockResolvedValueOnce(makeGiftDoc({ coinValue: 10 })) // gift
-      .mockResolvedValueOnce(makeUserDoc({ shyCoins: 500 })) // sender
-      .mockResolvedValueOnce(makeBackpackDoc(1)) // backpack has exactly 1
-      .mockResolvedValueOnce(makeUserDoc({ displayName: 'Bob' })) // recipient
-      .mockResolvedValueOnce(ECONOMY_CONFIG_DOC) // config
-      .mockResolvedValueOnce({ exists: false }) // giftWall
-      .mockResolvedValueOnce({ exists: false }) // giftRankings
-      .mockResolvedValue({ exists: false });
-
-    // Transaction should delete the backpack item since 1 - 1 = 0
-    const txDeleteMock = jest.fn();
-    mockRunTransaction.mockImplementation(async (cb) => {
-      const mockTx = {
-        get: jest.fn().mockResolvedValue({ exists: true, data: () => ({ quantity: 1 }) }),
-        update: jest.fn(),
-        delete: txDeleteMock,
-      };
-      return cb(mockTx);
-    });
-
-    const app = createApp('user-A');
-    const res = await request(app)
-      .post('/api/economy/gift-batch')
-      .send({ recipientIds: ['user-B'], giftId: 'gift-rose', quantity: 1, fromBackpack: true });
-
-    expect(res.status).toBe(200);
-    expect(txDeleteMock).toHaveBeenCalled();
-  });
-});
 
 // ═══════════════════════════════════════════════════════════════════
 // Purchase — production verification paths (lines 1272-1285)
