+++ express-api/tests/routes/identity-graph-write.test.js
@@ -929,117 +929,6 @@ describe('Admin auth — all graph routes', () => {
   });
 });
 
-describe('GET /api/admin/bans/check — validation and edge cases', () => {
-  test('returns 400 when no query parameters provided', async () => {
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check').expect(400);
-    expect(res.body.error).toMatch(/At least one identifier required/);
-  });
-
-  test('returns isBanned=false for expired suspension', async () => {
-    const expiredIdentifier = {
-      type: 'ip',
-      value: '1.2.3.4',
-      metadata: {},
-      addedAt: 1000,
-      source: 'login',
-      suspension: {
-        isActive: true,
-        level: 'full',
-        duration: '7d',
-        reason: 'Spam',
-        suspendedAt: 1000,
-        expiresAt: 1000, // long expired
-      },
-    };
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [makeGraphDoc('graph-1', { identifiers: [expiredIdentifier] })],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?ip=1.2.3.4').expect(200);
-    expect(res.body.isBanned).toBe(false);
-  });
-
-  test('returns isBanned=true for permanent suspension (expiresAt=null)', async () => {
-    const permanentIdentifier = {
-      type: 'uid',
-      value: '1001',
-      metadata: {},
-      addedAt: 1000,
-      source: 'login',
-      suspension: {
-        isActive: true,
-        level: 'full',
-        duration: 'permanent',
-        reason: 'Permanent ban',
-        suspendedAt: 1000,
-        expiresAt: null,
-      },
-    };
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [makeGraphDoc('graph-1', { identifiers: [permanentIdentifier] })],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?uid=1001').expect(200);
-    expect(res.body.isBanned).toBe(true);
-    expect(res.body.level).toBe('full');
-    expect(res.body.expiresAt).toBeNull();
-  });
-
-  test('matches fingerprint identifier correctly', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        makeGraphDoc('graph-1', {
-          identifiers: [makeSuspendedIdentifier('fingerprint', 'fp-match')],
-        }),
-      ],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?fingerprint=fp-match').expect(200);
-    expect(res.body.isBanned).toBe(true);
-  });
-
-  test('normalises IPv4-mapped IPv6 in ban check query', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        makeGraphDoc('graph-1', {
-          identifiers: [makeSuspendedIdentifier('ip', '5.6.7.8')],
-        }),
-      ],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?ip=::ffff:5.6.7.8').expect(200);
-    expect(res.body.isBanned).toBe(true);
-  });
-
-  test('does not match inactive suspension (suspension=null)', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        makeGraphDoc('graph-1', {
-          identifiers: [
-            {
-              type: 'ip',
-              value: '1.2.3.4',
-              metadata: {},
-              addedAt: 1000,
-              source: 'login',
-              suspension: null,
-            },
-          ],
-        }),
-      ],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?ip=1.2.3.4').expect(200);
-    expect(res.body.isBanned).toBe(false);
-  });
-});
-
 describe('PUT /api/admin/bans/graph/:id — unsuspend all', () => {
   test('unsuspend all identifiers clears all suspensions', async () => {
     mockDocGet.mockImplementation((path) => {
@@ -1179,65 +1068,3 @@ describe('Error handling — 500 responses', () => {
 // ═══════════════════════════════════════════════════════════════
 // Security: Query param type coercion (CodeQL type confusion fix)
 // ═══════════════════════════════════════════════════════════════
-
-describe('GET /api/admin/bans/check — array query param coercion', () => {
-  test('handles duplicate ip params (array) by using first value', async () => {
-    // Express parses ?ip=1.2.3.4&ip=5.6.7.8 as ip: ['1.2.3.4', '5.6.7.8']
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        makeGraphDoc('graph-1', {
-          identifiers: [makeSuspendedIdentifier('ip', '1.2.3.4')],
-        }),
-      ],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?ip=1.2.3.4&ip=5.6.7.8').expect(200);
-    // Should use the first value and find the ban
-    expect(res.body.isBanned).toBe(true);
-  });
-
-  test('handles duplicate fingerprint params (array) by using first value', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        makeGraphDoc('graph-1', {
-          identifiers: [makeSuspendedIdentifier('fingerprint', 'fp-first')],
-        }),
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/admin/bans/check?fingerprint=fp-first&fingerprint=fp-second')
-      .expect(200);
-    expect(res.body.isBanned).toBe(true);
-  });
-
-  test('handles duplicate uid params (array) by using first value', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        makeGraphDoc('graph-1', {
-          identifiers: [makeSuspendedIdentifier('uid', '1001')],
-        }),
-      ],
-    });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?uid=1001&uid=2002').expect(200);
-    expect(res.body.isBanned).toBe(true);
-  });
-
-  test('returns 400 when all params are empty after coercion', async () => {
-    const app = createApp();
-    // No params at all
-    const res = await request(app).get('/api/admin/bans/check').expect(400);
-    expect(res.body.error).toMatch(/At least one identifier required/);
-  });
-
-  test('handles single string ip param normally (no coercion needed)', async () => {
-    mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [] });
-    const app = createApp();
-    const res = await request(app).get('/api/admin/bans/check?ip=9.9.9.9').expect(200);
-    expect(res.body.isBanned).toBe(false);
-  });
-});
