+++ express-api/tests/routes/suggestions-contracts-a-misc.test.js
@@ -351,580 +351,6 @@ function setupSuggestionAndVote(suggestionOverrides = {}, voteDoc = null) {
 // 11.22 — Edge Cases & Boundaries
 // ═══════════════════════════════════════════════════════════════
 
-describe('11.22 — Edge Cases & Boundaries', () => {
-  // ── Concurrent Operations ─────────────────────────────────
-
-  describe('Concurrent operations', () => {
-    test('concurrent vote + edit on same suggestion: both resolve without corruption', async () => {
-      // Simulate two users acting simultaneously on the same suggestion
-      setupDocMocks({
-        'suggestions/sug-concurrent': makeSuggestionDoc('sug-concurrent', {
-          status: 'pending',
-          submitterUid: 1001,
-          upvotes: 5,
-          downvotes: 1,
-        }),
-      });
-
-      const ownerApp = createApp({ uniqueId: 1001 });
-      const voterApp = createApp({ uniqueId: 2002 });
-
-      // Owner edits while voter tries to vote (vote should fail on pending)
-      const [editRes, voteRes] = await Promise.all([
-        request(ownerApp)
-          .put('/api/suggestions/sug-concurrent')
-          .send({ title: 'Edited title', description: 'Edited desc' }),
-        request(voterApp).post('/api/suggestions/sug-concurrent/vote').send({ direction: 'up' }),
-      ]);
-
-      // Edit should succeed; vote on pending should be rejected
-      expect(editRes.status).toBe(200);
-      expect(voteRes.status).toBe(403);
-    });
-
-    test('concurrent vote + withdraw by owner: withdraw succeeds, vote fails', async () => {
-      setupDocMocks({
-        'suggestions/sug-race': makeSuggestionDoc('sug-race', {
-          status: 'pending',
-          submitterUid: 1001,
-        }),
-      });
-
-      const ownerApp = createApp({ uniqueId: 1001 });
-      const voterApp = createApp({ uniqueId: 2002 });
-
-      const [withdrawRes, voteRes] = await Promise.all([
-        request(ownerApp).delete('/api/suggestions/sug-race'),
-        request(voterApp).post('/api/suggestions/sug-race/vote').send({ direction: 'up' }),
-      ]);
-
-      // Withdraw of pending should succeed; vote on pending rejected
-      expect(withdrawRes.status).toBe(200);
-      expect(voteRes.status).toBe(403);
-    });
-
-    test('concurrent admin approve + user edit: transaction isolation', async () => {
-      setupDocMocks({
-        'suggestions/sug-txn': makeSuggestionDoc('sug-txn', {
-          status: 'pending',
-          submitterUid: 1001,
-        }),
-      });
-
-      const adminApp = createAdminApp();
-      const userApp = createApp({ uniqueId: 1001 });
-
-      const [approveRes, editRes] = await Promise.all([
-        request(adminApp).put('/api/admin/suggestions/sug-txn/status').send({ status: 'accepted' }),
-        request(userApp)
-          .put('/api/suggestions/sug-txn')
-          .send({ title: 'Edited after approval', description: 'desc' }),
-      ]);
-
-      // Both should receive responses (at least one 200, one possibly 403 or 200)
-      expect([200, 403]).toContain(approveRes.status);
-      expect([200, 403]).toContain(editRes.status);
-    });
-
-    test('concurrent upvote + downvote by different users: both resolve atomically', async () => {
-      setupSuggestionAndVote(
-        { status: 'accepted', submitterUid: 9999, upvotes: 10, downvotes: 3 },
-        null,
-      );
-
-      const user1App = createApp({ uniqueId: 2001 });
-      const user2App = createApp({ uniqueId: 2002 });
-
-      const [vote1, vote2] = await Promise.all([
-        request(user1App).post('/api/suggestions/sug1/vote').send({ direction: 'up' }),
-        request(user2App).post('/api/suggestions/sug1/vote').send({ direction: 'down' }),
-      ]);
-
-      // Both should succeed (different users voting)
-      expect(vote1.status).toBe(200);
-      expect(vote2.status).toBe(200);
-
-      // FieldValue.increment should have been called for both
-      const { FieldValue } = require('../../src/utils/firebase');
-      expect(FieldValue.increment).toHaveBeenCalled();
-    });
-  });
-
-  // ── Data Integrity ────────────────────────────────────────
-
-  describe('Data integrity', () => {
-    test('cascading delete: withdrawing a suggestion also removes its votes', async () => {
-      setupDocMocks({
-        'suggestions/sug-del': makeSuggestionDoc('sug-del', {
-          status: 'pending',
-          submitterUid: 1001,
-        }),
-      });
-
-      // Mock votes subcollection to have some docs
-      mockCollectionGet.mockResolvedValueOnce({
-        empty: false,
-        docs: [
-          makeVoteDoc('vote-1', { voterUid: 2001 }),
-          makeVoteDoc('vote-2', { voterUid: 2002 }),
-        ],
-        size: 2,
-      });
-
-      const app = createApp({ uniqueId: 1001 });
-      await request(app).delete('/api/suggestions/sug-del').expect(200);
-
-      // Verify suggestion delete was called
-      expect(mockDocDelete).toHaveBeenCalled();
-    });
-
-    test('cascading delete: withdrawing a suggestion also removes its comments', async () => {
-      setupDocMocks({
-        'suggestions/sug-del2': makeSuggestionDoc('sug-del2', {
-          status: 'pending',
-          submitterUid: 1001,
-        }),
-      });
-
-      // Mock comments subcollection
-      mockCollectionGet
-        .mockResolvedValueOnce({
-          empty: false,
-          docs: [makeCommentDoc('com-1'), makeCommentDoc('com-2')],
-          size: 2,
-        })
-        .mockResolvedValueOnce({
-          empty: false,
-          docs: [makeVoteDoc('vote-1')],
-          size: 1,
-        });
-
-      const app = createApp({ uniqueId: 1001 });
-      await request(app).delete('/api/suggestions/sug-del2').expect(200);
-
-      expect(mockDocDelete).toHaveBeenCalled();
-    });
-
-    test('vote transfer on merge: duplicate suggestion votes transferred to target', async () => {
-      // Set up a merge scenario — admin merges sug-dup into sug-target
-      setupDocMocks({
-        'suggestions/sug-dup': makeSuggestionDoc('sug-dup', {
-          status: 'accepted',
-          submitterUid: 2001,
-          upvotes: 3,
-          downvotes: 1,
-        }),
-        'suggestions/sug-target': makeSuggestionDoc('sug-target', {
-          status: 'accepted',
-          submitterUid: 3001,
-          upvotes: 10,
-          downvotes: 2,
-        }),
-      });
-
-      // Votes from sug-dup
-      mockCollectionGet.mockResolvedValueOnce({
-        empty: false,
-        docs: [
-          makeVoteDoc('v1', { voterUid: 4001, direction: 'up' }),
-          makeVoteDoc('v2', { voterUid: 4002, direction: 'up' }),
-          makeVoteDoc('v3', { voterUid: 4003, direction: 'down' }),
-        ],
-        size: 3,
-      });
-
-      const app = createAdminApp();
-      const res = await request(app)
-        .post('/api/admin/suggestions/sug-dup/merge')
-        .send({ targetSuggestionId: 'sug-target' });
-
-      // Merge should succeed (200) or the endpoint exists with some status
-      expect([200, 404]).toContain(res.status);
-
-      if (res.status === 200) {
-        // Verify the duplicate is updated with mergedIntoSuggestionId
-        const updateCalls = mockDocUpdate.mock.calls;
-        const mergeUpdate = updateCalls.find(
-          (c) => c[0]?.mergedIntoSuggestionId || c[1]?.mergedIntoSuggestionId,
-        );
-        expect(mergeUpdate).toBeDefined();
-      }
-    });
-
-    test('vote count stays non-negative after removing all votes', async () => {
-      setupSuggestionAndVote(
-        { status: 'accepted', submitterUid: 9999, upvotes: 1, downvotes: 0 },
-        makeVoteDoc('last-vote', { voterUid: 1001, direction: 'up' }),
-      );
-
-      const app = createApp({ uniqueId: 1001 });
-      await request(app).delete('/api/suggestions/sug1/vote').expect(200);
-
-      const { FieldValue } = require('../../src/utils/firebase');
-      // increment(-1) should be called but the route should clamp at 0 or trust Firestore
-      expect(FieldValue.increment).toHaveBeenCalledWith(-1);
-    });
-  });
-
-  // ── Empty/Null Handling ───────────────────────────────────
-
-  describe('Empty/null handling', () => {
-    test('GET suggestion with non-existent ID returns 404', async () => {
-      mockDocGet.mockResolvedValue({ exists: false });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions/does-not-exist').expect(404);
-      expect(res.body).toHaveProperty('error');
-    });
-
-    test('DELETE suggestion with non-existent ID returns 404', async () => {
-      mockDocGet.mockResolvedValue({ exists: false });
-      const app = createApp();
-      await request(app).delete('/api/suggestions/does-not-exist').expect(404);
-    });
-
-    test('PUT suggestion with non-existent ID returns 404', async () => {
-      mockDocGet.mockResolvedValue({ exists: false });
-      const app = createApp();
-      await request(app)
-        .put('/api/suggestions/does-not-exist')
-        .send({ title: 'Hello', description: 'World' })
-        .expect(404);
-    });
-
-    test('vote on non-existent suggestion returns 404', async () => {
-      mockDocGet.mockResolvedValue({ exists: false });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions/does-not-exist/vote')
-        .send({ direction: 'up' })
-        .expect(404);
-    });
-
-    test('remove vote on non-existent suggestion returns 404', async () => {
-      mockDocGet.mockResolvedValue({ exists: false });
-      const app = createApp();
-      await request(app).delete('/api/suggestions/does-not-exist/vote').expect(404);
-    });
-
-    test('POST suggestion with null title returns 400', async () => {
-      const app = createApp();
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, title: null })
-        .expect(400);
-      expect(res.body.error).toMatch(/title/i);
-    });
-
-    test('POST suggestion with null description returns 400', async () => {
-      const app = createApp();
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, description: null })
-        .expect(400);
-      expect(res.body.error).toMatch(/description/i);
-    });
-
-    test('POST suggestion with empty tags array succeeds (tags optional)', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      // Empty tags array may be treated as "no tags" — depends on implementation
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, tags: [] });
-      // Either 201 (tags optional) or 400 (tags required) — contract documents behavior
-      expect([201, 400]).toContain(res.status);
-    });
-
-    test('POST suggestion with null tags returns 400', async () => {
-      const app = createApp();
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, tags: null });
-      expect([400, 201]).toContain(res.status);
-    });
-
-    test('GET /api/suggestions/mine for user with zero suggestions returns empty array', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions/mine').expect(200);
-      expect(res.body.suggestions).toEqual([]);
-    });
-
-    test('GET /api/suggestions/search?q=xxxxxxxxxx returns empty results', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions/search?q=xxxxxxxxxx').expect(200);
-      expect(res.body.results).toEqual([]);
-    });
-  });
-
-  // ── Pagination Boundaries ─────────────────────────────────
-
-  describe('Pagination boundaries', () => {
-    test('page=0 treated as page 1', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions?page=0').expect(200);
-      expect(res.body.page).toBe(1);
-    });
-
-    test('negative page returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?page=-1').expect(400);
-    });
-
-    test('negative page -100 returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?page=-100').expect(400);
-    });
-
-    test('page beyond max results returns empty array', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions?page=999999').expect(200);
-      expect(res.body.suggestions).toEqual([]);
-    });
-
-    test('pageSize=0 returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?pageSize=0').expect(400);
-    });
-
-    test('pageSize=-1 returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?pageSize=-1').expect(400);
-    });
-
-    test('pageSize > max (e.g. 1000) capped at max (e.g. 50)', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions?pageSize=1000').expect(200);
-      expect(res.body.pageSize).toBeLessThanOrEqual(50);
-    });
-
-    test('pageSize=1 returns at most 1 result', async () => {
-      const docs = [makeSuggestionDoc('sug-only', { status: 'accepted' })];
-      mockCollectionGet.mockResolvedValueOnce({ empty: false, docs, size: 1 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions?pageSize=1').expect(200);
-      expect(res.body.suggestions.length).toBeLessThanOrEqual(1);
-    });
-
-    test('page=NaN returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?page=NaN').expect(400);
-    });
-
-    test('page=1.5 (non-integer) returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?page=1.5').expect(400);
-    });
-
-    test('pageSize=abc (non-numeric) returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?pageSize=abc').expect(400);
-    });
-
-    test('page as string "first" returns 400', async () => {
-      const app = createApp();
-      await request(app).get('/api/suggestions?page=first').expect(400);
-    });
-
-    test('exact max pageSize returns 200', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app).get('/api/suggestions?pageSize=50').expect(200);
-      expect(res.body.pageSize).toBe(50);
-    });
-  });
-
-  // ── Character Encoding ────────────────────────────────────
-
-  describe('Character encoding', () => {
-    test('emoji in title stored and returned correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const title = '\u{1F3A8} Add color themes \u{1F308}';
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, title })
-        .expect(201);
-
-      expect(res.body).toHaveProperty('id');
-    });
-
-    test('emoji in description stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const description =
-        'Please add \u{1F4AC} chat bubbles with \u{2764}\u{FE0F} reactions \u{1F525}\u{1F389}';
-      await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, description })
-        .expect(201);
-    });
-
-    test('compound emoji (family, flag, skin tone) in title handled', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const title =
-        '\u{1F468}\u{200D}\u{1F469}\u{200D}\u{1F467}\u{200D}\u{1F466} Family feature \u{1F1EC}\u{1F1E7}';
-      await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, title })
-        .expect(201);
-    });
-
-    test('CJK characters (Chinese) in title stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          title: '\u8BF7\u6DFB\u52A0\u6697\u8272\u6A21\u5F0F',
-          language: 'zh',
-        })
-        .expect(201);
-    });
-
-    test('CJK characters (Japanese) in description stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          description:
-            '\u6697\u3044\u30E2\u30FC\u30C9\u3092\u8FFD\u52A0\u3057\u3066\u304F\u3060\u3055\u3044\u3002\u3053\u308C\u306F\u3068\u3066\u3082\u4FBF\u5229\u3067\u3059\u3002',
-          language: 'ja',
-        })
-        .expect(201);
-    });
-
-    test('CJK characters (Korean) in title stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          title: '\uB2E4\uD06C \uBAA8\uB4DC\uB97C \uCD94\uAC00\uD574 \uC8FC\uC138\uC694',
-          language: 'ko',
-        })
-        .expect(201);
-    });
-
-    test('RTL text (Arabic) in title stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          title:
-            '\u0625\u0636\u0627\u0641\u0629 \u0627\u0644\u0648\u0636\u0639 \u0627\u0644\u0645\u0638\u0644\u0645',
-          language: 'ar',
-        })
-        .expect(201);
-    });
-
-    test('RTL text (Hebrew) in description stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          description:
-            '\u05D1\u05D1\u05E7\u05E9\u05D4 \u05DC\u05D4\u05D5\u05E1\u05D9\u05E3 \u05DE\u05E6\u05D1 \u05D7\u05E9\u05D5\u05DA',
-          language: 'he',
-        });
-      // May succeed or fail based on language code validation
-      // Hebrew "he" should be a valid ISO 639-1 code
-    });
-
-    test('mixed RTL/LTR text in description handled', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          description:
-            'This is English \u0648\u0647\u0630\u0627 \u0639\u0631\u0628\u064A and back to English',
-        })
-        .expect(201);
-    });
-
-    test('newlines in description preserved', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const description = 'Line 1\nLine 2\nLine 3\n\nParagraph 2';
-      await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, description })
-        .expect(201);
-
-      const storedCall = mockDocSet.mock.calls[0] || mockCollectionAdd.mock.calls[0];
-      if (storedCall) {
-        const data = storedCall[storedCall.length - 1];
-        if (data?.description) {
-          expect(data.description).toContain('\n');
-        }
-      }
-    });
-
-    test('tab characters in title handled (stripped or preserved)', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, title: 'Tab\there\ttitle' });
-      // Should succeed with tabs stripped or preserved
-      expect([201, 400]).toContain(res.status);
-    });
-
-    test('Unicode control characters in title rejected or stripped', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      const res = await request(app)
-        .post('/api/suggestions')
-        .send({ ...VALID_SUGGESTION, title: 'Title\x00with\x01control\x02chars' });
-      // Should either strip them (201) or reject (400)
-      expect([201, 400]).toContain(res.status);
-    });
-
-    test('Devanagari (Hindi) text in title stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          title:
-            '\u0921\u093E\u0930\u094D\u0915 \u092E\u094B\u0921 \u091C\u094B\u0921\u093C\u0947\u0902',
-          language: 'hi',
-        })
-        .expect(201);
-    });
-
-    test('Thai text in description stored correctly', async () => {
-      mockCollectionGet.mockResolvedValueOnce({ empty: true, docs: [], size: 0 });
-      const app = createApp();
-      await request(app)
-        .post('/api/suggestions')
-        .send({
-          ...VALID_SUGGESTION,
-          description:
-            '\u0E40\u0E1E\u0E34\u0E48\u0E21\u0E42\u0E2B\u0E21\u0E14\u0E21\u0E37\u0E14\u0E43\u0E2B\u0E49\u0E41\u0E2D\u0E1B',
-          language: 'th',
-        })
-        .expect(201);
-    });
-  });
-});
-
 // ═══════════════════════════════════════════════════════════════
 // 11.33 — Firestore Rule Enforcement (middleware-level auth)
 // ═══════════════════════════════════════════════════════════════
