+++ express-api/tests/routes/suggestions-duplicates-core.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 /**
  * Tests for suggestion duplicate detection, merge, dispute, admin search,
  * and blocked-topic similarity matching.
@@ -916,185 +917,3 @@ describe('GET /api/admin/suggestions — Admin search & duplicate highlighting',
 // ═══════════════════════════════════════════════════════════════
 // 11.79 — Blocked Topic Similarity Matching
 // ═══════════════════════════════════════════════════════════════
-
-describe('GET /api/suggestions/blocked?q= — Blocked topic check', () => {
-  test('exact title match blocked', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add cryptocurrency payments',
-            rejectReason: 'Against our values',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=Add+cryptocurrency+payments')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(true);
-  });
-
-  test('case-insensitive match blocked', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add Cryptocurrency Payments',
-            rejectReason: 'Against our values',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=add+cryptocurrency+payments')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(true);
-  });
-
-  test('high similarity match blocked (>80%)', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add cryptocurrency payments to the app',
-            rejectReason: 'Against our values',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=Add+cryptocurrency+payments+to+the+application')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(true);
-  });
-
-  test('low similarity passes', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add cryptocurrency payments',
-            rejectReason: 'Against our values',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=Add+dark+mode+to+profile')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(false);
-  });
-
-  test('returns rejection reason from original', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add cryptocurrency payments',
-            rejectReason: 'Against our values and policies',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=Add+cryptocurrency+payments')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(true);
-    expect(res.body).toHaveProperty('rejectReason', 'Against our values and policies');
-  });
-
-  test('returns link to original rejected suggestion', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add cryptocurrency payments',
-            rejectReason: 'Against our values',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=Add+cryptocurrency+payments')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(true);
-    expect(res.body).toHaveProperty('originalSuggestionId', 'sug-rejected-1');
-  });
-
-  test('multiple blocked topics can match, all shown', async () => {
-    mockCollectionGet.mockResolvedValueOnce({
-      empty: false,
-      docs: [
-        {
-          id: 'bt1',
-          data: () => ({
-            title: 'Add crypto payments',
-            rejectReason: 'Against policy',
-            originalSuggestionId: 'sug-rejected-1',
-          }),
-        },
-        {
-          id: 'bt2',
-          data: () => ({
-            title: 'Add cryptocurrency wallet',
-            rejectReason: 'Not feasible',
-            originalSuggestionId: 'sug-rejected-2',
-          }),
-        },
-        {
-          id: 'bt3',
-          data: () => ({
-            title: 'Cryptocurrency integration',
-            rejectReason: 'Legal concerns',
-            originalSuggestionId: 'sug-rejected-3',
-          }),
-        },
-      ],
-    });
-    const app = createApp();
-    const res = await request(app)
-      .get('/api/suggestions/blocked?q=Add+cryptocurrency+payments')
-      .expect(200);
-
-    expect(res.body.blocked).toBe(true);
-    expect(res.body).toHaveProperty('matches');
-    expect(Array.isArray(res.body.matches)).toBe(true);
-    expect(res.body.matches.length).toBeGreaterThanOrEqual(2);
-    // Each match should include reason and link
-    res.body.matches.forEach((match) => {
-      expect(match).toHaveProperty('rejectReason');
-      expect(match).toHaveProperty('originalSuggestionId');
-    });
-  });
-});
